requirements.txt
mcp[cli]>=1.2.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.30.0,<3.0.0